refactor: extract folder api handlers
